Move classes only relevant for ferdigstilling to corresponding interface.
