+++ web/src/main/kotlin/no/nav/su/se/bakover/web/routes/drift/FixIverksettingerResponseJson.kt
@@ -1,6 +1,6 @@
 package no.nav.su.se.bakover.web.routes.drift
 
-import no.nav.su.se.bakover.service.søknadsbehandling.OpprettManglendeJournalpostOgBrevdistribusjonResultat
+import no.nav.su.se.bakover.service.søknadsbehandling.FerdigstillSøknadsbehandingIverksettingService
 
 data class FixIverksettingerResponseJson(
     val journalposteringer: Journalposteringer,
@@ -35,7 +35,7 @@ data class FixIverksettingerResponseJson(
     )
 
     companion object {
-        fun OpprettManglendeJournalpostOgBrevdistribusjonResultat.toJson() = FixIverksettingerResponseJson(
+        fun FerdigstillSøknadsbehandingIverksettingService.OpprettManglendeJournalpostOgBrevdistribusjonResultat.toJson() = FixIverksettingerResponseJson(
             journalposteringer = Journalposteringer(
                 ok = this.journalpostresultat.mapNotNull { it.orNull() }.map {
                     Journalpost(
